+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- 2021-02-28 Add wp, equivalent of std::weak_ptr.
 - 2021-02-18 Add Thread::sleep(). Make TimeTick a class. 
 - 2021-02-18 Change the implementation of `sp<NcArray>`'s operator[].
 - 2021-02-18 Add `synchroized()` and `ManualResetEvent`.